소수점 2째자리까지 표현하게 수정, 테스트 코드 수정

+++ com.codescroll.widget/src/main/java/com/codescroll/widget/chart/CSPieChart.java
@@ -54,7 +54,7 @@ public class CSPieChart extends Composite{
 		titleLabel.setText(title);
 	}
 	
-	public void setValue(int value){
+	public void setValue(Float value){
 		checkWidget();
 		pie.setValue(value); 
 	}
